examples/requirements.txt
matplotlib>=3.2.1
seaborn>=0.10.0
scipy>=1.4.1
Pillow>=7.0.0
numpy>=1.15.4
pandas>=0.25.3
tqdm>=4.45.0
